db: Initial events ms backend
