--- events/code/controllers/events.json
@@ -1,28 +1,0 @@
-{
-  "events": [
-    {
-      "id": 1,
-      "name": "Event 1",
-      "date": "2022-10-01",
-      "location": "New York City"
-    },
-    {
-      "id": 2,
-      "name": "Event 2",
-      "date": "2022-11-15",
-      "location": "Los Angeles"
-    },
-    {
-      "id": 3,
-      "name": "Event 3",
-      "date": "2022-12-31",
-      "location": "Chicago"
-    },
-    {
-      "id": 4,
-      "name": "New Event",
-      "date": "2023-01-15",
-      "location": "San Francisco"
-    }
-  ]
-}
