--- events/code/middleware/errorHandler.js
@@ -1,11 +1,0 @@
-export function ErrorHandler(err, req, res, next) {
-  console.log('Middleware Error Handling');
-  const errStatus = err.statusCode || 500;
-  const errMsg = err.message || 'Something went wrong';
-  res.status(errStatus).json({
-    success: false,
-    status: errStatus,
-    message: errMsg,
-    stack: process.env.NODE_ENV === 'development' ? err.stack : {},
-  });
-}
