--- events/code/middleware/exampleMiddleware.js
@@ -1,4 +1,0 @@
-export async function checkName(req, res, next){
-  console.log('I do not know you');
-  next();
-}
