--- events/code/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "ex1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "./node_modules/nodemon/bin/nodemon.js start.js",
-    "dev win": "./node_modules/.bin/nodemon start.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.2",
-    "express": "^4.17.1"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.1"
-  }
-}
