--- events/code/variables.env
@@ -1,3 +1,0 @@
-NODE_ENV=development
-PORT=3010
-HOST=127.0.0.1
